add option 'ssh-options' to pass options to ssh
From: Lars Wirzenius <liw@liw.fi>
Date: Sat, 23 Feb 2013 18:53:12 +0000
Subject: Re: [PATCH] add option 'ssh-options' to add ssh command line
options

"
This requires that Obnam parses the options so that it can pass
them onto ssh. Such parsing is either quite complicated (when it
emulates the way the shell does it), or prone to be surprising
to users. Your patch is of the latter kind, and will cause
surprises when the ssh options need to include spaces.

If you rework the patch to use cliapp's string list setting
type, the user can give each ssh option as a separate Obnam
option, and embedded spaces will work. This would be a much
cleaner solution.

Such options also only affect the SFTP transfers when an
underlying ssh binary is used, but not when a pure Paramiko
is used. However, I can live with that.

As an aside, for bandwidth limiting, trickle(1) seems to be a nice,
generic option.
"