Try to restore group even when not root.
Normal users can't set the uid to a user other than themselves, but
they can set the group id if they are in that group.  Previously it
was ignoring uid and gid if it wasn't root, update to not change the
uid if not root and try to set the gid, ignoring any errors.

This also makes some improvements to the metadata_tests.  Make sure
magic uid and gid isn't the current user.  Expand fake_lchown to use
-1 as don't update that UID for the above group restore tests to pass.
Create oserror_lchown for code coverage to exercise the error path.
test_does_not_set_uid_when_not_running_as_root was using fake_lchown,
but checking the actual file, which would normally be owned by the
user that just created it, so it wasn't an effective test.
test_does_not_set_gid_when_not_running_as_root no longer applies, as
it does try to set the group.  Add a test for lchown set failure, and
a test to verify it actually sets the file to every group the user is
in.