support --to location in verify
This works just like the --to option in restore, only it checks it
against the repository files.  This allows restoring to a different
directroy, and verifying the restore, which is also useful if the
backup directory is mounted in a different location, which could be
because the system was booted from a rescue image.