store and restore character and block special device node numbers
These special devices were restored with 0, 0 as the major and minor
numbers causing them to be unusable.  Without which a system backup
and restore would leave the system unbootable.  The major and minor
numbers are saved at the end of the metadata so as to make the
repository format backwards compatible.