improve locking reporting
Let the user know when the backup is being blocked by a lock as it is
useful to warn the user as soon as possible that there might be a
problem, especially when the back just started otherwise they might go
on to something else thinking it is working.

Use unix line endings for the lock file, include the date, and print
the contents when a lock timeout happens.  Having the date is useful
to have an idea just how long the lock has been sitting there and if
it is safe to force remove it.