Don't leave locks around when interrupted.
If backup or fsck were interrupted with Control-C or a program
exception the repository locks would be left in place.  In this change
the lock manager keeps track of all out standing locks and fs.close()
will call into the lock manager to cleanup those locks, which is safe
as long as there are no more writes.  This catches exceptions in
backup and fsck plugins to call fs.close().  Idealy there would never
be any locks to cleanup, but if interrupted there are.

Not cleaning up the locks means automated backups are most likely
blocked, adn running force-lock can be dangerous because if another
client is actively backing up it will continue, but without the locks
it put in place.