close VirtualFileSystem before closing sftp
For an easy demonstration of problems, back up to an sftp repository,
but then give an invalid backup directory, without the lockmgr updates
it would have left locks, with the lockmgr changes it crashes still
leaving the locks, because the force_unlock_all was being called
after the sftp repository was closed so it couldn't clean up the
locks.  Moved it to before.